CodeSmell: Api/background cleanup (#4147)
* background apis

* moved to background

* moved to background

* moved to background

* movied to backgrounds

* moving all run/dashboard to background URLs

refactor the code for smipler code

* moving some functions into static

* fix path to timerjobs
