--- src/api/routes/system/system-timerjobs.php
@@ -1,7 +1,0 @@
-<?php
-
-$app->group('/timerjobs', function () {
-    $this->post('/run', function () {
-        $this->SystemService->runTimerJobs();
-    });
-});
