+++ src/api/routes/system/system.php
@@ -1,6 +1,5 @@
 <?php
 
-use ChurchCRM\Utils\LoggerUtils;
 use Slim\Http\Request;
 use Slim\Http\Response;
 use ChurchCRM\Service\NotificationService;
@@ -8,16 +7,8 @@ use ChurchCRM\dto\Notification\UiNotification;
 use ChurchCRM\Service\TaskService;
 
 $app->group('/system', function () {
-    $this->post('/csp-report', 'logCSPReportAPI');
     $this->get('/notification', 'getUiNotificationAPI');
 });
-
-function logCSPReportAPI(Request $request, Response $response, array $args)
-{
-    $input = json_decode($request->getBody());
-    $log = json_encode($input, JSON_PRETTY_PRINT);
-    LoggerUtils::getAppLogger()->warn($log);
-}
 
 function getUiNotificationAPI(Request $request, Response $response, array $args)
 {
